requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-filter==2.2.0
django-storages==1.8
djangorestframework==3.11.0
google-api-core==1.16.0
google-auth==1.10.1
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
multidict==4.7.4
Pillow==7.0.0
protobuf==3.11.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
yarl==1.4.2
